removed unused controller entries in dashadmin narrowed user dependent access to only admin, later changes to come dash and home controllers uneffected, except dashadmin all tested

+++ WebProject/Controllers/OrderDetailController.cs
@@ -18,6 +18,7 @@ namespace WebProject.Controllers
         // GET: OrderDetail
         public ActionResult Index(int? product, int? order)
         {
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             var order_Details = db.Order_Details.Include(o => o.Order).Include(o => o.Product);
             if (product != null)
                 order_Details = order_Details.Where(t => t.ProductID == product);
@@ -29,11 +30,8 @@ namespace WebProject.Controllers
         // GET: OrderDetail/Details/5
         public ActionResult Details(int? order, int? product)
         {
-            if (order == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            if (product == null)
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
+            if (product == null || order == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
@@ -48,6 +46,7 @@ namespace WebProject.Controllers
         // GET: OrderDetail/Create
         public ActionResult Create()
         {
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             ViewBag.OrderID = new SelectList(db.Orders, "OrderID", "CustomerID");
             ViewBag.ProductID = new SelectList(db.Products, "ProductID", "ProductName");
             return View();
@@ -60,6 +59,7 @@ namespace WebProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "OrderID,ProductID,UnitPrice,Quantity,Discount")] Order_Detail order_Detail)
         {
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             if (ModelState.IsValid)
             {
                 db.Order_Details.Add(order_Detail);
@@ -75,11 +75,8 @@ namespace WebProject.Controllers
         // GET: OrderDetail/Edit/5
         public ActionResult Edit(int? order, int? product)
         {
-            if (order == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            if (product == null)
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
+            if (product == null || order == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
@@ -100,6 +97,7 @@ namespace WebProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "OrderID,ProductID,UnitPrice,Quantity,Discount")] Order_Detail order_Detail)
         {
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             if (ModelState.IsValid)
             {
                 db.Entry(order_Detail).State = EntityState.Modified;
@@ -114,11 +112,8 @@ namespace WebProject.Controllers
         // GET: OrderDetail/Delete/5
         public ActionResult Delete(int? order, int? product)
         {
-            if (order == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            if (product == null)
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
+            if (product == null || order == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
@@ -135,6 +130,7 @@ namespace WebProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int? order, int? product)
         {
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             Order_Detail order_Detail = db.Order_Details.Find(order, product);
             db.Order_Details.Remove(order_Detail);
             db.SaveChanges();
