+++ WebProject/Controllers/TerritoriesController.cs
@@ -18,6 +18,7 @@ namespace WebProject.Controllers
         // GET: Territories
         public ActionResult Index(int? region)
         {
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             var territories = db.Territories.Include(t => t.Region);
             if (region != null)
                 territories = territories.Where(t => t.RegionID == region);
@@ -27,6 +28,7 @@ namespace WebProject.Controllers
         // GET: Territories/Details/5
         public ActionResult Details(string id)
         {
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -42,6 +44,7 @@ namespace WebProject.Controllers
         // GET: Territories/Create
         public ActionResult Create()
         {
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             ViewBag.RegionID = new SelectList(db.Regions, "RegionID", "RegionDescription");
             return View();
         }
@@ -53,6 +56,7 @@ namespace WebProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "TerritoryID,TerritoryDescription,RegionID")] Territory territory)
         {
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             if (ModelState.IsValid)
             {
                 db.Territories.Add(territory);
@@ -67,6 +71,7 @@ namespace WebProject.Controllers
         // GET: Territories/Edit/5
         public ActionResult Edit(string id)
         {
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -87,6 +92,7 @@ namespace WebProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "TerritoryID,TerritoryDescription,RegionID")] Territory territory)
         {
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             if (ModelState.IsValid)
             {
                 db.Entry(territory).State = EntityState.Modified;
@@ -100,6 +106,7 @@ namespace WebProject.Controllers
         // GET: Territories/Delete/5
         public ActionResult Delete(string id)
         {
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -117,6 +124,7 @@ namespace WebProject.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(string id)
         {
+            if (Session["a"] == null) return new HttpStatusCodeResult(HttpStatusCode.Unauthorized);
             Territory territory = db.Territories.Find(id);
             db.Territories.Remove(territory);
             db.SaveChanges();
